Use a submodule for picovoice porcupine
This gives us access to the header files, and with some C++
magic we can make sure that the functions we load from the library
have the right types.

With that, we can realize that the 2.0 interface is different,
which explains segfaults. At the same time, the keyword files
we have are tagged 1.9.0, so use the latest porcupine commit
in 1.9.0